requirements.txt
astroid==2.2.5
isort==4.3.20
lazy-object-proxy==1.4.1
lxml==4.3.4
mccabe==0.6.1
pylint==2.3.1
python-docx==0.8.10
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2
